ActiveKids_Project 2/requirements.txt
flask
pymongo
redis
cassandra-driver
neo4j
python-dotenv
